Initial developments (simple GARCH realization)
Simple GARCH model is realized in two possible ways. 

In the first method each step in simple GARCH model is coded with several functions (the code is based on developments from one of quantitative study courses). Parameters for GARCH model are calculated through maximization of likelihood function (SciPy library is being used in the process). 

The second method uses ARCH library.

Next steps will include modification of the code with inclusion of elements, responsible for Markov-Switching part of the model.